+++ backend/nest_backend/src/user/services/user/user.service.ts
@@ -6,6 +6,7 @@ import { CreateUserDto } from 'src/dtos/createUser.dto';
 import { FriendshipDto } from 'src/dtos/friendship.dto';
 import { ShortResponseUserDto } from 'src/dtos/shortResponseUser.dto';
 import { GameMode, StatusMode } from 'src/entities/user.entity';
+import { UpdateUserDto } from 'src/dtos/updateUser.dto';
 
 @Injectable()
 export class UserService {
@@ -14,6 +15,9 @@ export class UserService {
   ) {}
 
   createUser(createUserDto: CreateUserDto) {
+    if (!createUserDto) {
+      throw new Error("CreateUserDto is undefined!");
+    }
     const newUser = this.userRepository.create(createUserDto);
     newUser.status = StatusMode.ONLINE;
     newUser.blacklist = [];
@@ -30,6 +34,9 @@ export class UserService {
   }
 
   findUserById(id: string) {
+    if (!id) {
+      throw new Error("Id is undefined!");
+    }
     return this.userRepository.findOne({ where: { id: id } });
   }
   getUsers() {
@@ -37,20 +44,30 @@ export class UserService {
   }
 
   async addFriend(friendshipDto: FriendshipDto) {
+    if (!friendshipDto) {
+      throw new Error("friendshipDto is undefined!");
+    }
     const user = await this.findUserById(friendshipDto.userId);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     const friend = await this.findUserById(friendshipDto.friendId);
+    if (friend == null) {
+      throw new Error("friend not found!")
+    }
     user.friends = await this.findFriends(friendshipDto.userId);
     if (user.friends) {
       user.friends.push(friend);
     } else {
       user.friends = [friend];
     }
-    try {
-      await this.userRepository.save(user);
-    } catch (e) {}
+    await this.userRepository.save(user);
   }
 
   async deleteFriend(friendshipDto: FriendshipDto) {
+    if (!friendshipDto) {
+      throw new Error("FriendshipDto is undefined!");
+    }
     return await this.userRepository.query(
       ` DELETE FROM users_friends_users u
         WHERE (u."usersId_1" = $1 AND u."usersId_2" = $2)
@@ -60,6 +77,9 @@ export class UserService {
   }
 
   async findFriends(id: string): Promise<User[]> {
+    if (!id) {
+      throw new Error("id is undefined!");
+    }
     return await this.userRepository.query(
       ` SELECT * 
         FROM users U
@@ -75,6 +95,9 @@ export class UserService {
   }
 
   async findFriendsDto(id: string): Promise<ShortResponseUserDto[]> {
+    if (!id) {
+      throw new Error("id is undefined!");
+    }
     return await this.userRepository.query(
       ` SELECT id, username, status
         FROM users U
@@ -89,6 +112,9 @@ export class UserService {
     );
   }
   async findUserByLogin(login: string) {
+    if (!login) {
+      throw new Error("login is undefined!");
+    }
     const user = await this.userRepository
       .createQueryBuilder('user')
       .where('user.login = :login', { login: login })
@@ -100,6 +126,9 @@ export class UserService {
   }
 
   async findUserIdByUsername(username: string) {
+    if (!username) {
+      throw new Error("Username is undefined!");
+    }
     const user = await this.userRepository
       .createQueryBuilder('user')
       .where('user.username = :username', { username: username })
@@ -113,6 +142,9 @@ export class UserService {
   }
 
   async findUsernameById(id: string) {
+    if (!id) {
+      throw new Error("id is undefined!");
+    }
     const user = await this.userRepository
       .createQueryBuilder('user')
       .where('user.id = :id', { id: id })
@@ -122,30 +154,64 @@ export class UserService {
     }
     return user.username;
   }
-  async updateUserPicture(id: string, image: string) {
-    const user = await this.findUserById(id);
-    user.image = image;
+  async updateUser(updateUserDto: UpdateUserDto) {
+    if (!updateUserDto) {
+      throw new Error("UpdateUserDto is undefined!");
+    }
+    const user = await this.findUserById(updateUserDto.id);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
+    if (user.username != updateUserDto.username) {
+      const anotherUser = await this.findUserIdByUsername(updateUserDto.username);
+      if (anotherUser.user != null) {
+        throw new Error('Account with this username already exists.');
+      }
+    }
+    user.image = updateUserDto.image;
+    user.username = updateUserDto.username;
     return this.userRepository.save(user);
   }
   async addToBlacklist(userId: string, blackId: string) {
+    if (!userId || !blackId) {
+      throw new Error("userId or blackId is undefined!");
+    }
     const user = await this.findUserById(userId);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     if (!user.blacklist.includes(blackId)) {
       user.blacklist.push(blackId);
     }
     return this.userRepository.save(user);
   }
   async deleteFromBlacklist(userId: string, blackId: string) {
+    if (!userId || !blackId) {
+      throw new Error("userId or blackId is undefined!");
+    }
     const user = await this.findUserById(userId);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     if (user.blacklist.includes(blackId)) {
       user.blacklist = user.blacklist.filter((id) => id != blackId);
     }
     return this.userRepository.save(user);
   }
   async checkBlacklist(userId: string, checkId: string) {
+    if (!userId || !checkId) {
+      throw new Error("userId or checkId is undefined")
+    }
     const user = await this.findUserById(userId);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     return user.blacklist.includes(checkId);
   }
   async getShortInfoById(id: string) {
+    if (!id) {
+      throw new Error("id is undefined!");
+    }
     return {
         id: id,
         username: await this.findUsernameById(id),
@@ -153,6 +219,9 @@ export class UserService {
     };
   }
   async getShortInfoByIds(ids: string[]) {
+    if (!ids) {
+      throw new Error("ids are undefined!");
+    }
     const res = [];
     for (const id of ids) {
       res.push(await this.getShortInfoById(id));
@@ -160,26 +229,56 @@ export class UserService {
     return res;
   }
   async setTwoFactorAuthSecret(secret: string, login: string) {
+    if (!secret || !login) {
+      throw new Error("secret code or login is undefined!");
+    }
     const user = await this.findUserByLogin(login);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     user.twoFactorAuthSecret = secret;
     return this.userRepository.save(user);
   }
   async switchTwoFactorAuth(login: string, condition: boolean) {
+    if (!login || !condition) {
+      throw new Error("login or condition is undefined!");
+    }
     const user = await this.findUserByLogin(login);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     user.twoFactorAuthIsEnabled = condition;
     return this.userRepository.save(user);
   }
   async changeUserStatus(userId: string, status: StatusMode) {
+    if (!userId || !status) {
+      throw new Error("userId or status is undefined!")
+    }
     const user = await this.findUserById(userId);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     user.status = status;
     return this.userRepository.save(user);
   }
   async findStatusById(userId: string) {
+    if (!userId) {
+      throw new Error("userId wasn't found!");
+    }
     const user = await this.findUserById(userId);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     return user.status;
   }
   async changeGameMode(userId: string, mode: GameMode) {
+    if (!userId || !mode) {
+      throw new Error("userId or gameMode is undefined!");
+    }
     const user = await this.findUserById(userId);
+    if (user == null) {
+      throw new Error("user not found!");
+    }
     user.preferredGameMode = mode;
     return this.userRepository.save(user);
   }
